fix(callback) Refactor code to use more clear naming
Make sure files implementing a callback are ending with `_callback`. Also have the abstract class named Interface to make it more clear what it's purpose is.
